+++ source/random-spec.ts
@@ -4,6 +4,6 @@ describe('R.random', () => {
   it('happy', () => {
     const result = random(1, 4)
 
-    result // $ExpectType number[]
+    result // $ExpectType number
   })
 })